chore: Update npm dependencies and imports
This commit updates the npm dependencies in the `package.json` file and adds new imports for the `axios` and `cheerio` packages. The `axios` package will be used for making HTTP requests, while the `cheerio` package will be used for parsing HTML. These changes ensure that the project has the latest dependencies and enables the use of these new packages.